Moved data iterator out of notifications page, still some more cleanup to go
